# api-wellezy/storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "API Wellezy",
        "description": "Esta es la documentación de la API de Wellezy.",
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://127.0.0.1:8000/"
        },
        {
            "url": "http://localhost"
        }
    ],
    "paths": {
        "/api/airports/{code}": {
            "post": {
                "tags": [
                    "Airports"
                ],
                "summary": "Busca aeropuertos por el nombre de la ciudad.",
                "description": "Obtiene información de aeropuertos según el código de ciudad proporcionado.",
                "operationId": "480f90aa698cba7b4edd97425a83cf78",
                "parameters": [
                    {
                        "name": "code",
                        "in": "path",
                        "description": "Nombre de la ciudad para buscar aeropuertos",
                        "required": true,
                        "schema": {
                            "type": "string"
                        },
                        "example": "Medellin"
                    }
                ],
                "requestBody": {
                    "description": "Nombre de la ciudad",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "code": {
                                        "type": "string",
                                        "example": "Medellin"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Aeropuertos encontrados",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "properties": {
                                            "id": {
                                                "type": "number",
                                                "example": 1
                                            },
                                            "city": {
                                                "type": "string",
                                                "example": "Medellin"
                                            },
                                            "name": {
                                                "type": "string",
                                                "example": "Enrique Olaya Herrera"
                                            },
                                            "country": {
                                                "type": "string",
                                                "example": "Colombia"
                                            },
                                            "iata": {
                                                "type": "string",
                                                "example": "EOH"
                                            },
                                            "created_at": {
                                                "type": "string",
                                                "example": "2023-08-30T14:37:13.000000Z"
                                            },
                                            "updated_at": {
                                                "type": "string",
                                                "example": "2023-08-30T14:37:13.000000Z"
                                            }
                                        },
                                        "type": "object"
                                    }
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "No se encontraron aeropuertos",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "No airports found in this city"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "Error en la solicitud",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Error fetching data: [detalle del error]"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/flights": {
            "post": {
                "tags": [
                    "Flights"
                ],
                "summary": "Busca vuelos según criterios proporcionados",
                "description": "Busca vuelos basados en la ciudad de salida, llegada, número de pasajeros y hora de salida.",
                "operationId": "ea47c22ac381c72c304eff4cbb63f3ed",
                "requestBody": {
                    "description": "Datos para buscar vuelos",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "searchs": {
                                        "type": "integer",
                                        "example": 250
                                    },
                                    "qtyPassengers": {
                                        "type": "integer",
                                        "example": 1
                                    },
                                    "adult": {
                                        "type": "integer",
                                        "example": 1
                                    },
                                    "itinerary": {
                                        "type": "array",
                                        "items": {
                                            "properties": {
                                                "departureCity": {
                                                    "type": "string",
                                                    "example": "mde"
                                                },
                                                "arrivalCity": {
                                                    "type": "string",
                                                    "example": "lhr"
                                                },
                                                "hour": {
                                                    "type": "string",
                                                    "format": "date-time",
                                                    "example": "2024-10-09T00:00:00+0000"
                                                }
                                            },
                                            "type": "object"
                                        }
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Vuelos encontrados",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "properties": {
                                            "dateOfDeparture": {
                                                "type": "string",
                                                "format": "date",
                                                "example": "2024-02-09"
                                            },
                                            "timeOfDeparture": {
                                                "type": "string",
                                                "example": "08:00"
                                            },
                                            "dateOfArrival": {
                                                "type": "string",
                                                "format": "date",
                                                "example": "2024-02-09"
                                            },
                                            "timeOfArrival": {
                                                "type": "string",
                                                "example": "10:00"
                                            },
                                            "marketingCarrier": {
                                                "type": "string",
                                                "example": "AV"
                                            },
                                            "flightOrtrainNumber": {
                                                "type": "string",
                                                "example": "AV123"
                                            },
                                            "locationId": {
                                                "properties": {
                                                    "departureCity": {
                                                        "type": "string",
                                                        "example": "MDE"
                                                    },
                                                    "arrivalCity": {
                                                        "type": "string",
                                                        "example": "LHR"
                                                    }
                                                },
                                                "type": "object"
                                            }
                                        },
                                        "type": "object"
                                    }
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "Error en la solicitud",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Error fetching data: [detalle del error]"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
...
}